tts-service/requirements.txt
# TTS Service 依赖

# Web 框架
fastapi==0.116.1
uvicorn[standard]==0.35.0
pydantic==2.11.10

# HTTP 客户端
httpx==0.28.1

# ModelScope 和 TTS 相关
modelscope==1.33.0
kantts==1.0.1
ttsfrd==0.2.1

# 数值计算
numpy==1.26.4
scipy==1.10.1

# 工具库
simplejson==3.20.2
addict==2.4.0
sortedcontainers==2.4.0
tqdm==4.67.1
PyYAML==6.0.3
requests==2.32.5
tensorboardX==2.6.4
pytorch-wavelets==1.3.0
soundfile==0.13.1
sox==1.5.0
datasets==3.6.0

